num_segments configuration

+++ kafka-client/src/main/java/dev/responsive/kafka/internal/utils/StoreUtil.java
@@ -47,7 +47,7 @@ public final class StoreUtil {
     }
   }
 
-  public static long computeSegmentInterval(final long retentionPeriod, final int numSegments) {
+  public static long computeSegmentInterval(final long retentionPeriod, final long numSegments) {
     return retentionPeriod / numSegments;
   }
